requirements.txt
cycler==0.11.0
kiwisolver==1.3.1
matplotlib==3.3.4
numpy==1.18.5
opencv-python==4.4.0.40
Pillow==7.2.0
pygame==1.9.6
pyparsing==3.0.9
python-dateutil==2.8.2
six==1.16.0
